+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 6.0.0 [28-01-2025]
+**Updated** Category class 
+Removed to template handling from within the Category class.
+Bump to 6.0.0, as the previous update was a major
+
 ### 5.8.0 [28-01-2025]
 **Added** Category class 
 A generic class for processing discourse categories and the topics they contain